requirements.txt
openai
langchain
langchain-community
langchain-openai
langgraph
langgraph-checkpoint-postgres
psycopg[binary]
psycopg2-binary
azure-identity 
azure-keyvault-secrets
azure-search-documents
azure-ai-formrecognizer
SQLAlchemy
google-auth
fastapi
uvicorn
itsdangerous
pytest